+++ src/main/java/com/mcb/creditfactory/service/car/CarAdapter.java
@@ -18,7 +18,7 @@ public class CarAdapter implements CollateralObject {
     }
 
     @Override
-    public Short getYear() {
+    public Integer getYear() {
         return car.getYear();
     }
